feat: Add hardware interface links (PTY, SocketCAN, TAP)
- Introduce `SocketCanLink`, `TapLink`, and `PtyLink` for hardware communication
- Guard hardware link headers with `HAS_HARDWARE` macro
- Add `WIREBIT_HARDWARE` CMake option to control hardware link compilation
- Update Makefile to pass `HARDWARE` flag to CMake and XMake
- Add `hardware` option to xmake.lua to enable hardware interface links
- Define `HAS_HARDWARE` macro when hardware option is enabled in CMake and XMake
- Expand buffer sizes in `can_bus_hub.cpp` and `can_broadcast.cpp` examples for node naming
- Cast `timeout_ms` to `uint64_t` in `can_node.cpp` to prevent comparison warnings
- Change `values` array type to `unsigned int` in `string_to_mac` to match `sscanf` behavior
- Remove redundant `DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN` from test files

+++ examples/can_node.cpp
@@ -125,7 +125,7 @@ int mode_recv(int node_id, int timeout_ms = 5000) {
     while (true) {
         // Check timeout
         uint64_t elapsed_ms = (now_ns() - start_time) / 1000000;
-        if (elapsed_ms > timeout_ms) {
+        if (elapsed_ms > static_cast<uint64_t>(timeout_ms)) {
             echo::warn("Timeout reached").yellow();
             break;
         }
